Move SiGe equilibrium structure calculator to the src module.
